backend/requirements.txt
# Django Core
Django>=5.0.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# Security & Configuration
python-decouple>=3.8
django-cors-headers>=4.3.1

# Image Processing (for product images)
Pillow>=10.0.0

# Production Server
gunicorn>=21.0.0